version 0.0.23

+++ setup.py
@@ -1,14 +1,15 @@
-import setuptools
+import os,setuptools
 
-with open("version.txt", "r") as f:
-    version = f.read()
-    if version:
-        versionno = version.split("=")[-1].strip("' ")
-        arr = [int(item) for item in versionno.split(".")]
-        arr[2] += 1
-        version = "%d.%d.%d" % tuple(arr)
-        with open("version.txt", "w") as w:
-            w.write(version)
+if os.path.isfile("version.txt"):
+    with open("version.txt", "r") as f:
+        version = f.read()
+        if version:
+            versionno = version.split("=")[-1].strip("' ")
+            arr = [int(item) for item in versionno.split(".")]
+            arr[2] += 1
+            version = "%d.%d.%d" % tuple(arr)
+            with open("version.txt", "w") as w:
+                w.write(version)
 
 setuptools.setup(
     name="gecosistema_learning",
